requirements.txt
timm==0.4.12
transformers==4.12.5
ruamel_yaml
opencv-python
scikit-image
matplotlib
pycocotools
pycocoevalcap
datasets
sentencepiece

accelerate
sklearn
yacs
ninja
einops
ftfy
scikit-learn
h5py